I cleaned the code up a bit
What i did was basically just putting the main algorithm inside a function that I called conclusions(). As an input I decided to create (value) that does everything that result was doing in the original code. In the body I called the function entering as value "result", so it works the same as yours.
ringraziame